Project_Graduation_2/requirements.txt
# Coca-Cola Sorting System - Python Dependencies

# Computer Vision & AI
opencv-python>=4.8.0
Pillow>=10.0.0
ultralytics>=8.0.0  # YOLOv8
numpy>=1.24.0

# Serial Communication
pyserial>=3.5

# Database
# sqlite3 is included in Python standard library

# GUI
# tkinter is included in Python standard library

# Additional utilities for YOLOv8
torch>=2.0.0  # PyTorch for YOLOv8
torchvision>=0.15.0
